+++ lib/rabarber.rb
@@ -8,12 +8,13 @@ require "active_record"
 require "active_support"
 
 require_relative "rabarber/input/base"
-require_relative "rabarber/input/actions"
-require_relative "rabarber/input/dynamic_rules"
+require_relative "rabarber/input/action"
+require_relative "rabarber/input/dynamic_rule"
+require_relative "rabarber/input/role"
 require_relative "rabarber/input/roles"
-require_relative "rabarber/input/types/booleans"
-require_relative "rabarber/input/types/procs"
-require_relative "rabarber/input/types/symbols"
+require_relative "rabarber/input/types/boolean"
+require_relative "rabarber/input/types/proc"
+require_relative "rabarber/input/types/symbol"
 
 require_relative "rabarber/missing/base"
 require_relative "rabarber/missing/actions"
@@ -25,7 +26,8 @@ require_relative "rabarber/controllers/concerns/authorization"
 require_relative "rabarber/helpers/helpers"
 require_relative "rabarber/models/concerns/has_roles"
 require_relative "rabarber/models/role"
-require_relative "rabarber/permissions"
+
+require_relative "rabarber/core/permissions"
 
 require_relative "rabarber/railtie"