+++ lib/rabarber/input/base.rb
@@ -5,7 +5,7 @@ module Rabarber
     class Base
       attr_reader :value, :error_type, :error_message
 
-      def initialize(value, error_type, error_message)
+      def initialize(value, error_type = Rabarber::InvalidArgumentError, error_message = default_error_message)
         @value = value
         @error_type = error_type
         @error_message = error_message
@@ -15,13 +15,17 @@ module Rabarber
         valid? ? processed_value : raise_error
       end
 
-      private
-
       def valid?
         raise NotImplementedError
       end
 
+      private
+
       def processed_value
+        raise NotImplementedError
+      end
+
+      def default_error_message
         raise NotImplementedError
       end