+++ lib/rabarber/input/types/boolean.rb
@@ -3,15 +3,19 @@
 module Rabarber
   module Input
     module Types
-      class Booleans < Rabarber::Input::Base
-        private
-
+      class Boolean < Rabarber::Input::Base
         def valid?
           [true, false].include?(value)
         end
 
+        private
+
         def processed_value
           value
+        end
+
+        def default_error_message
+          "Value must be a Boolean"
         end
       end
     end