+++ lib/rabarber/input/types/proc.rb
@@ -3,15 +3,19 @@
 module Rabarber
   module Input
     module Types
-      class Procs < Rabarber::Input::Base
-        private
-
+      class Proc < Rabarber::Input::Base
         def valid?
-          value.is_a?(Proc)
+          value.is_a?(::Proc)
         end
+
+        private
 
         def processed_value
           value
+        end
+
+        def default_error_message
+          "Value must be a Proc"
         end
       end
     end