+++ spec/rabarber/missing/actions_spec.rb
@@ -7,13 +7,13 @@ RSpec.describe Rabarber::Missing::Actions do
 
   before { allow(Rabarber::Configuration.instance).to receive(:when_actions_missing).and_return(callable_double) }
 
-  after { Rabarber::Permissions.action_rules.delete(DummyAuthController) }
+  after { Rabarber::Core::Permissions.action_rules.delete(DummyAuthController) }
 
   context "when controller is not specified" do
     let(:controller) { nil }
 
     context "when action is missing" do
-      before { Rabarber::Permissions.add(DummyAuthController, :index, [:admin], nil, nil) }
+      before { Rabarber::Core::Permissions.add(DummyAuthController, :index, [:admin], nil, nil) }
 
       it "calls configuration" do
         expect(callable_double).to receive(:call).with([:index], { controller: DummyAuthController })
@@ -33,7 +33,7 @@ RSpec.describe Rabarber::Missing::Actions do
     let(:controller) { DummyAuthController }
 
     context "when action is missing" do
-      before { Rabarber::Permissions.add(DummyAuthController, :index, [:admin], nil, nil) }
+      before { Rabarber::Core::Permissions.add(DummyAuthController, :index, [:admin], nil, nil) }
 
       it "calls configuration" do
         expect(callable_double).to receive(:call).with([:index], { controller: DummyAuthController })
@@ -49,9 +49,9 @@ RSpec.describe Rabarber::Missing::Actions do
     end
 
     context "when action is missing in another controller" do
-      before { Rabarber::Permissions.add(DummyController, :non_existent, [:admin], nil, nil) }
+      before { Rabarber::Core::Permissions.add(DummyController, :non_existent, [:admin], nil, nil) }
 
-      after { Rabarber::Permissions.action_rules[DummyController].pop }
+      after { Rabarber::Core::Permissions.action_rules[DummyController].pop }
 
       it "does not call configuration" do
         expect(callable_double).not_to receive(:call)