+++ spec/rabarber/models/concerns/has_roles_spec.rb
@@ -320,4 +320,10 @@ RSpec.describe Rabarber::HasRoles do
       it { is_expected.to be_empty }
     end
   end
+
+  describe "#roleable_class" do
+    subject { described_class.roleable_class }
+
+    it { is_expected.to eq(User) }
+  end
 end